requirements.txt
absl-py==2.1.0
accelerate==1.4.0
annotated-types==0.7.0
anyio==4.8.0
cabocha-python==0.69.3
certifi==2025.1.31
charset-normalizer==3.4.1
chex==0.1.82
click==8.1.8
distro==1.9.0
etils==1.12.0
exceptiongroup==1.2.2
fastapi==0.115.8
filelock==3.17.0
flax==0.7.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.29.1
idna==3.10
importlib_resources==6.5.2
jax==0.4.13
jaxlib==0.4.13
Jinja2==3.1.5
jiter==0.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mecab==0.996.5
mecab-python3==1.0.10
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==1.26.4
openai==1.64.0
opt_einsum==3.4.0
optax==0.1.4
orbax-checkpoint==0.5.16
packaging==24.2
protobuf==5.29.3
psutil==7.0.0
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-config==0.1.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
safetensors==0.5.2
scipy==1.12.0
sentencepiece==0.2.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tensorstore==0.1.72
tokenizers==0.21.0
toolz==1.0.0
torch==2.6.0
tqdm==4.67.1
transformers==4.49.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
zipp==3.21.0
